more fixes for java 1.1 api spec compliance, assisted by japitools

+++ libraries/javalib/java/awt/AWTEvent.java
@@ -35,7 +35,7 @@ public class AWTEvent
 	final public static long ADJUSTMENT_EVENT_MASK = 0x100;
 	final public static long ITEM_EVENT_MASK = 0x200;
 	final public static long TEXT_EVENT_MASK = 0x400;
-	final public static long RESERVED_ID_MAX = 1999;
+	final public static int RESERVED_ID_MAX = 1999;
 	final static long DISABLED_MASK = 0x80000000;
 	final static long TEMP_DISABLED_MASK = 0x40000000;
 	static Component keyTgt;
